refactor(admin): extract RK9 Players sub-view to its own component
PlayersView owns its search/sort/expansion state + query + virtualized table;
parent renders <PlayersView active refreshKey> in the players branch. No behavior
change.
