+++ apps/web/src/components/admin/external-data.tsx
@@ -5,17 +5,12 @@ import { useSearchParams } from "next/navigation";
 import { useVirtualizer } from "@tanstack/react-virtual";
 import { toast } from "sonner";
 import {
-  ArrowDown,
-  ArrowUp,
-  ArrowUpDown,
   ChevronDown,
   ChevronRight,
   ExternalLink,
   ListFilter,
-  Search,
 } from "lucide-react";
 import { Badge } from "@/components/ui/badge";
-import { Input } from "@/components/ui/input";
 import { Skeleton } from "@/components/ui/skeleton";
 import { cn } from "@/lib/utils";
 import { useSupabaseQuery } from "@/lib/supabase";
@@ -53,7 +48,6 @@ import { ExternalDataToolbar } from "./external-data-toolbar";
 import { SelectionBar } from "./external-data-selection-bar";
 import { EventList } from "./external-data-cards";
 import { ExpandedRowData } from "./expanded-row-data";
-import { PlayerExpandedData } from "./player-expanded-data";
 import { QueueStrip } from "./external-data-queue-strip";
 import { ExternalDataFilters } from "./external-data-filters";
 import { StatusBadge } from "./external-data-status-badge";
@@ -64,6 +58,7 @@ import {
   compareValues,
   type SortState,
 } from "./external-data-table-helpers";
+import { PlayersView } from "./external-data-players-view";
 
 // Sentinel for render-time tab-change reset (avoids useEffect for derived state)
 const UNINITIALIZED = Symbol();
@@ -1156,101 +1151,12 @@ export function ExternalData() {
   // RK9 sub-view: events or players
   const [rk9View, setRk9View] = useState<"events" | "players">("events");
 
-  // Player search
-  const [playerSearch, setPlayerSearch] = useState("");
-
-  // RK9 players query — always called (Rules of Hooks), but skips fetch when not on players view
-  const { data: rk9Players, isLoading: rk9PlayersLoading } = useSupabaseQuery(
-    /* istanbul ignore next */
-    async (sb) => {
-      if (rk9View !== "players") return [];
-      const { data, error } = await sb
-        .schema("rk9")
-        .from("players")
-        .select(
-          "id, player_id_masked, first_name, last_name, country, standings(count)"
-        )
-        .order("last_name", { ascending: true });
-      if (error) throw error;
-      return (data ?? []) as unknown as Array<{
-        id: number;
-        player_id_masked: string | null;
-        first_name: string;
-        last_name: string;
-        country: string;
-        standings: [{ count: number }];
-      }>;
-    },
-    [refreshKey, rk9View === "players" ? "load" : "skip"]
-  );
-
-  // Client-side player filter
-  const filteredPlayers = (rk9Players ?? []).filter((p) => {
-    if (!playerSearch) return true;
-    const q = playerSearch.toLowerCase();
-    return (
-      p.first_name?.toLowerCase().includes(q) ||
-      p.last_name?.toLowerCase().includes(q) ||
-      p.player_id_masked?.toLowerCase().includes(q)
-    );
-  });
-
-  // Expanded player row state
-  const [expandedPlayerId, setExpandedPlayerId] = useState<number | null>(null);
-
   const scrollRef = useRef<HTMLDivElement>(null);
-  const playerScrollRef = useRef<HTMLDivElement>(null);
-
-  type PlayerSortCol = "name" | "id" | "country" | "events";
-  const [playerSort, setPlayerSort] = useState<{
-    column: PlayerSortCol;
-    direction: "asc" | "desc";
-  }>({ column: "name", direction: "asc" });
-
-  function togglePlayerSort(col: PlayerSortCol) {
-    setPlayerSort((prev) =>
-      prev.column === col
-        ? { column: col, direction: prev.direction === "asc" ? "desc" : "asc" }
-        : { column: col, direction: "asc" }
-    );
-  }
-
-  const sortedPlayers = [...filteredPlayers].sort((a, b) => {
-    const dir = playerSort.direction === "asc" ? 1 : -1;
-    switch (playerSort.column) {
-      case "name":
-        return (
-          `${a.first_name} ${a.last_name}`.localeCompare(
-            `${b.first_name} ${b.last_name}`
-          ) * dir
-        );
-      case "id":
-        return (
-          (a.player_id_masked ?? "").localeCompare(b.player_id_masked ?? "") *
-          dir
-        );
-      case "country":
-        return (a.country ?? "").localeCompare(b.country ?? "") * dir;
-      case "events":
-        return (
-          ((a.standings[0]?.count ?? 0) - (b.standings[0]?.count ?? 0)) * dir
-        );
-      default:
-        return 0;
-    }
-  });
 
   const rowVirtualizer = useVirtualizer({
     count: currentRows.length,
     getScrollElement: () => scrollRef.current,
     estimateSize: () => 56,
-    overscan: 10,
-  });
-
-  const playerVirtualizer = useVirtualizer({
-    count: sortedPlayers.length,
-    getScrollElement: () => playerScrollRef.current,
-    estimateSize: () => 52,
     overscan: 10,
   });
 
@@ -1363,17 +1269,12 @@ export function ExternalData() {
           </div>
         )}
 
-        {/* Players search input — only in RK9 players view */}
+        {/* RK9 Players sub-view (search + table) vs the shared events toolbar */}
         {filters.source === "rk9" && rk9View === "players" ? (
-          <div className="relative">
-            <Search className="text-muted-foreground absolute top-1/2 left-2.5 h-3.5 w-3.5 -translate-y-1/2" />
-            <Input
-              placeholder="Search by name or player ID..."
-              value={playerSearch}
-              onChange={(e) => setPlayerSearch(e.target.value)}
-              className="h-8 pl-8 text-sm"
-            />
-          </div>
+          <PlayersView
+            active={filters.source === "rk9" && rk9View === "players"}
+            refreshKey={refreshKey}
+          />
         ) : (
           <>
             {/* Single toolbar — unified Sync + Import vocabulary */}
@@ -1530,148 +1431,6 @@ export function ExternalData() {
         <span className="text-muted-foreground text-xs">
           {bulkProgress.done}/{bulkProgress.total} — {bulkProgress.current}
         </span>
-      )}
-
-      {/* Players Table — only shown when source=rk9 + players view */}
-      {filters.source === "rk9" && rk9View === "players" && (
-        <div className="rounded-md border">
-          <div className="text-muted-foreground flex items-center gap-2 px-4 py-2 text-xs">
-            <ListFilter className="h-3.5 w-3.5" />
-            {rk9PlayersLoading
-              ? "Loading players…"
-              : `Showing ${filteredPlayers.length} players`}
-          </div>
-
-          {/* Players table header */}
-          <div
-            className="grid border-b"
-            style={{ gridTemplateColumns: "28px 1fr 120px 60px 60px" }}
-          >
-            <div className="h-10" />
-            {(["name", "id", "country", "events"] as const).map((col) => (
-              <div key={col} className="flex h-10 items-center px-2">
-                <button
-                  className="hover:text-foreground inline-flex items-center gap-1 text-xs font-medium whitespace-nowrap capitalize"
-                  onClick={() => togglePlayerSort(col)}
-                >
-                  {col === "id"
-                    ? "RK9 ID"
-                    : col.charAt(0).toUpperCase() + col.slice(1)}
-                  {playerSort.column === col ? (
-                    playerSort.direction === "asc" ? (
-                      <ArrowUp className="h-3 w-3" />
-                    ) : (
-                      <ArrowDown className="h-3 w-3" />
-                    )
-                  ) : (
-                    <ArrowUpDown className="h-3 w-3 opacity-40" />
-                  )}
-                </button>
-              </div>
-            ))}
-          </div>
-
-          {rk9PlayersLoading ? (
-            <div className="space-y-2 p-3">
-              {Array.from({ length: 6 }).map((_, i) => (
-                <Skeleton key={i} className="h-11 w-full" />
-              ))}
-            </div>
-          ) : filteredPlayers.length === 0 ? (
-            <div className="text-muted-foreground py-12 text-center text-sm">
-              {(rk9Players ?? []).length === 0
-                ? "No players found."
-                : "No players match your search."}
-            </div>
-          ) : (
-            /* Virtualized body */
-            <div
-              ref={playerScrollRef}
-              className="overflow-auto"
-              style={{ maxHeight: "calc(100vh - 300px)" }}
-            >
-              <div
-                style={{
-                  height: playerVirtualizer.getTotalSize(),
-                  position: "relative",
-                }}
-              >
-                {playerVirtualizer.getVirtualItems().map((virtualRow) => {
-                  const p = sortedPlayers[virtualRow.index];
-                  if (!p) return null;
-                  const isPlayerExpanded = expandedPlayerId === p.id;
-                  const eventCount = p.standings[0]?.count ?? 0;
-                  return (
-                    <div
-                      key={p.id}
-                      ref={playerVirtualizer.measureElement}
-                      data-index={virtualRow.index}
-                      className="border-b"
-                      style={{
-                        position: "absolute",
-                        top: 0,
-                        left: 0,
-                        width: "100%",
-                        transform: `translateY(${virtualRow.start}px)`,
-                      }}
-                    >
-                      <div
-                        className="hover:bg-muted/50 grid transition-colors"
-                        style={{
-                          gridTemplateColumns: "28px 1fr 120px 60px 60px",
-                        }}
-                      >
-                        {/* Chevron */}
-                        <div className="flex items-center justify-center py-3">
-                          <button
-                            className="hover:bg-muted flex h-5 w-5 items-center justify-center rounded"
-                            onClick={() =>
-                              setExpandedPlayerId(
-                                isPlayerExpanded ? null : p.id
-                              )
-                            }
-                            aria-label={
-                              isPlayerExpanded ? "Collapse" : "Expand"
-                            }
-                          >
-                            {isPlayerExpanded ? (
-                              <ChevronDown className="h-3.5 w-3.5" />
-                            ) : (
-                              <ChevronRight className="h-3.5 w-3.5" />
-                            )}
-                          </button>
-                        </div>
-                        {/* Name */}
-                        <div className="flex min-w-0 items-center px-3 py-3 text-xs">
-                          <span className="truncate">
-                            {[p.first_name, p.last_name]
-                              .filter(Boolean)
-                              .join(" ") || "—"}
-                          </span>
-                        </div>
-                        {/* Player ID */}
-                        <div className="text-muted-foreground flex items-center px-3 py-3 font-mono text-xs">
-                          {p.player_id_masked ?? "—"}
-                        </div>
-                        {/* Country */}
-                        <div className="flex items-center px-3 py-3 font-mono text-xs uppercase">
-                          {p.country ?? "—"}
-                        </div>
-                        {/* Events */}
-                        <div className="flex items-center px-3 py-3 text-xs">
-                          {eventCount}
-                        </div>
-                      </div>
-                      {isPlayerExpanded && (
-                        <PlayerExpandedData playerId={p.id} />
-                      )}
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          )}
-        </div>
       )}
 
       {/* Shared Events Table — hidden when source=rk9 and players view active */}
